% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agegroup_as_map.R
\name{agegroup_as_map}
\alias{agegroup_as_map}
\title{Categorize age into two yearly intervals, needed to apply MAP method}
\usage{
agegroup_as_map(data, age = "age")
}
\arguments{
\item{data}{processed data}

\item{age}{age of women}
}
\value{
\code{data.frame}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{agegroup_as_map()} converts \code{age} variable into a character vector named
\code{agegroup} with two yearly intervals between 14 and 50.
}
\examples{

## demonstrating using microdata
data("microdata")
## get only female
md <- subset(microdata, sex == 2)
## get those aged between 14 and 50
md <- subset(md, age >= 15 & age < 50)
## create age group into 2-yearly intervals
md <- agegroup_as_map(md, age = "age")

summary(md$agegroup)
table(md$agegroup)

}
