% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bangladesh}
\alias{bangladesh}
\title{Bangladesh 1974}
\format{
A data frame
}
\usage{
data(bangladesh)
}
\description{
The data gathered by the 1974 Bangladesh Retrospective Survey of Fertility and Mortality
can be used to demonstrate the estimation of child mortality from summary birth histories
using the Trussell version of the BRASS method and the Coale-Demeny model life tables
\code{\link{coale_demeny_ltm}}.
}
\details{
\itemize{
\item extracted from Display 6 on page 28 and Display 7 on page 29.
}
}
\references{
United Nations Population Studies (1990) Step-by-Step Guide
to the Estimation of Child Mortality No.107:1-83
(\href{https://www.un.org/en/development/desa/population/publications/pdf/mortality/stepguide_childmort.pdf}{United Nations})
}
\keyword{datasets}
