% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coef_mac_re}
\alias{coef_mac_re}
\title{Random effect coefficients for the Maternal Age Cohort-derived method}
\format{
A data frame
}
\source{
\href{https://journals.plos.org/plosmedicine/article?id=10.1371/journal.pmed.1000253}{PLoS MEDICINE}
}
\usage{
data(coef_mac_re)
}
\description{
Random effect coefficients provided by Rajaratnam et al for estimation of under-five child
mortality data from the maternal age Cohort-derived method (MAC).
}
\references{
Rajaratnam JK, Tran LN, Lopez AD, Murray CJL (2010) Measuring Under-Five Mortality: Validation of New Low-Cost Methods. PLOS Medicine 7(4): e1000253.
(\doi{10.1371/journal.pmed.1000253}{10.1371/journal.pmed.1000253})
}
\keyword{datasets}
