% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deathdays_distribution}
\alias{deathdays_distribution}
\title{Distribution of death days for calculating maternal age period-derived U5MR estimates}
\format{
A data frame
}
\source{
\href{https://journals.plos.org/plosmedicine/article?id=10.1371/journal.pmed.1000253}{PLoS MEDICINE}
}
\usage{
data(deathdays_distribution)
}
\description{
The data of distribution of birthdays for different regions.
}
\details{
\itemize{
\item \code{ASIA}
\item \code{LATC} (Latin America and the Carribbean)
\item \code{NAME} (North Africa and Middle East)
\item \code{SASE} (Sub-Saharan Africa, South/East)
\item \code{SAWC} (Sub-Saharan Africa, West/Central)
}
}
\references{
Rajaratnam JK, Tran LN, Lopez AD, Murray CJL (2010) Measuring Under-Five Mortality: Validation of New Low-Cost Methods. PLOS Medicine 7(4): e1000253.
(\doi{10.1371/journal.pmed.1000253}{10.1371/journal.pmed.1000253})
}
\keyword{datasets}
