% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microdata}
\alias{microdata}
\title{Fake data for Cambodia}
\format{
A data frame
}
\source{
\href{https://journals.plos.org/plosmedicine/article?id=10.1371/journal.pmed.1000253}{PLoS MEDICINE}
}
\usage{
data(microdata)
}
\description{
Fake data used to demonstrate the application of Cohort-derived and Period-derived methods
developed by Rajaratnam et al in 2010.
}
\details{
\code{iso3} - the iso3 code of the country from which your microdata come
\code{region} - the region that the country belongs in
\code{country} - name of the country
\code{svy_wt} - sample weight given to the respondent. If no sample weights are provided, then
generate this variable with a value of 1 for each respondent
\code{age} - age of the respondent in years: or time since first birth of the respondent in years
\code{sex} - sex of the respondent where 1 indicates male and 2 is female.
\code{ceb} - number of children ever born
\code{cd} - number of children that died
}
\references{
Rajaratnam JK, Tran LN, Lopez AD, Murray CJL (2010) Measuring Under-Five Mortality: Validation of New Low-Cost Methods. PLOS Medicine 7(4): e1000253.
(\doi{10.1371/journal.pmed.1000253}{10.1371/journal.pmed.1000253})
}
\keyword{datasets}
