% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panama}
\alias{panama}
\title{Panama 1976}
\format{
A data frame
}
\source{
\href{https://www.un.org/en/development/desa/population/publications/manual/estimate/demographic-estimation.asp}{United Nations Population Division}
}
\usage{
data(panama)
}
\description{
The data gathered by a survey in Panama between August and October 1976
can be used to demonstrate the estimation of child mortality from summary birth histories
using the Trussell version of the BRASS method and the Coale-Demeny model life tables
\code{\link{coale_demeny_ltm}}.
}
\details{
\itemize{
\item extracted from Table49 on page 78.
}
}
\references{
United Nations (1983) Manual X: indirect techniques for demographic
estimation. Population studies No. 81. New York: United Nations Department
of International Economic and Social Affairs
(\href{https://unstats.un.org/unsd/demographic/standmeth/handbooks/Manual_X-en.pdf}{United Nations})
}
\keyword{datasets}
