% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u5mr_period.R
\name{u5mr_period}
\alias{u5mr_period}
\title{Estimating under-five mortality using Maternal Age Period-derived method (MAP)}
\usage{
u5mr_period(
  data,
  child_born = "child_born",
  child_dead = "child_dead",
  agegrp = "agegrp",
  svy_wt = "svy_wt",
  iso3 = "KHM",
  svy_region = "ASIA",
  svy_year = 1234
)
}
\arguments{
\item{data}{preprocessed data}

\item{child_born}{children ever born}

\item{child_dead}{children dead}

\item{agegrp}{age grouping or time since first birth}

\item{svy_wt}{sample weights: if not available, use 1.}

\item{iso3}{the \code{iso3} code of the country from which the survey data come}

\item{svy_region}{region of the country from which the survey data come
\itemize{
\item \code{ASIA}
\item \code{LATC} (Latin America and the Caribbean)
\item \code{NAME} (North Africa and Middle East)
\item \code{SASE} (Sub-Saharan Africa, South/East)
\item \code{SAWC} (Sub-Saharan Africa, West/Central)
}}

\item{svy_year}{end of the survey}
}
\value{
\code{data.frame}
\itemize{
\item \code{ref_time}  - time between survey year and interpolated year
\item \code{year}  - reference year
\item \code{q5} - under-five mortality
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{u5mr_period()} uses the maternal age period-derived method (MAP) through summary
birth history information and maternal age (or time since first birth) to
calculate the under-five mortality estimates.
}
\details{
In this period-derived method, under-five mortality and reference time are estimated
through distributions of child birthdays and death days for different categories
of mothers, stratified by maternal information such as region, age, and number
of child ever born or dead. These distributions are used to find the expected number
of children ever born and dead in every year prior to the survey (up to 25 years) for
a mother in each particular strata.

By applying these distributions to each mother in each strata, and summing
across all strata, expected numbers of children ever born (CEB) and child dead
(CD) are generated for each year prior to the survey. The ratio of CD and
CEB for each year can then be calculated.

\strong{Computational Procedure}

The formulas used to quantify MAP method is as follows:

logit(\out{5q0<sub>tjk</sub>}) = \eqn{\beta}\out{<sup>0</sup><sub>t</sub>} +
\out{U<sub>tj</sub>} + \eqn{\beta}\out{<sup>1</sup><sub>t</sub>} x
logit(\out{CD<sub>tjk</sub>} / \out{CEB<sub>tjk</sub>}) +
\eqn{\epsilon}\out{<sub>tjk</sub>}

where

t = index of calendar time \eqn{\epsilon} (0, 24)
j = country
k = survey
\out{CD<sub>i</sub>} = total dead children in time bin \code{t}
\out{CEB<sub>i</sub>} = total children ever born in time bin \code{t}
}
\examples{

data("microdata")
## get only female
md <- subset(microdata, sex == 2)
## get those aged between 14 and 50
md <- subset(md, age >= 15 & age < 50)
## create age group into 2-yearly intervals
md <- agegroup_as_map(md, age = "age")
u5mr_period(md, child_born = "ceb", child_dead = "cd", agegrp = "agegroup",
            svy_wt = "svy_wt", iso3 = "KHM",
            svy_region = "ASIA", svy_year = 1234)

}
\references{
Rajaratnam JK, Tran LN, Lopez AD, Murray CJL (2010) Measuring Under-Five Mortality: Validation of New Low-Cost Methods. PLOS Medicine 7(4): e1000253.
(\doi{10.1371/journal.pmed.1000253}{10.1371/journal.pmed.1000253})
}
